+++ src/FuelClient.cc
@@ -156,6 +156,11 @@ class ignition::fuel_tools::FuelClientPrivate
 
   /// \brief Regex to parse Ignition Fuel world file URLs.
   public: std::unique_ptr<std::regex> urlWorldFileRegex;
+
+  /// \brief The set of licenses where the key is the name of the license
+  /// and the value is the license ID on a Fuel server. See the
+  /// PopulateLicenses function.
+  public: std::map<std::string, unsigned int> licenses;
 };
 
 //////////////////////////////////////////////////
@@ -417,10 +422,57 @@ Result FuelClient::UploadModel(const std::string &_pathToModelDir,
   // \todo(nkoenig) The ign-fuelserver expects an integer number for the
   // license information. The fuelserver should be modified to accept
   // a string. Otherwise, we have to bake into each client a mapping of
-  // license name to integer. For now, we are making a model
-  // "Creative Commons - Public Domain
-  // if (meta.has_legal()) {....}
-  form.emplace("license", "1");
+  // license name to integer.
+  //
+  // If we have legal, then attempt to fill in the correct license information.
+  if (meta.has_legal())
+  {
+    // Attempt to retrieve the available licenses, if we have no available
+    // licenses.
+    if (this->dataPtr->licenses.empty())
+    {
+      this->PopulateLicenses(_id.Server());
+      // Fail if a license has been requested, but we couldn't get the
+      // available licenses.
+      if (this->dataPtr->licenses.empty())
+      {
+        return Result(ResultType::UPLOAD_ERROR);
+      }
+    }
+
+    // Find the license by name.
+    std::map<std::string, unsigned int>::const_iterator licenseIt =
+      this->dataPtr->licenses.find(meta.legal().license());
+    if (licenseIt != this->dataPtr->licenses.end())
+    {
+      form.emplace("license", std::to_string(licenseIt->second));
+    }
+    // No license found, print an error and return.
+    else
+    {
+      std::string validLicenseNames;
+      auto end = this->dataPtr->licenses.end();
+      std::advance(end, -1);
+      for (licenseIt = this->dataPtr->licenses.begin();
+           licenseIt != end; ++licenseIt)
+      {
+        validLicenseNames += "    " + licenseIt->first + "\n";
+      }
+      validLicenseNames += "    " + licenseIt->first;
+
+      ignerr << "Invalid license[" << meta.legal().license() << "].\n"
+             << "  Valid licenses include:\n"
+             << validLicenseNames << std::endl;
+
+      return Result(ResultType::UPLOAD_ERROR);
+    }
+  }
+  // If there is no license information, then default to
+  // "Creative Commons - Public Domain"
+  else
+  {
+    form.emplace("license", "1");
+  }
 
   // Add tags
   std::string tags;
@@ -1159,3 +1211,23 @@ void FuelClientPrivate::AllFiles(const std::string &_path,
   }
 }
 
+
+//////////////////////////////////////////////////
+void FuelClient::PopulateLicenses(const ServerConfig &_server)
+{
+  ignition::fuel_tools::Rest rest;
+  RestResponse resp;
+
+  // Send the request.
+  resp = rest.Request(HttpMethod::GET, _server.Url().Str(),
+      _server.Version(), "licenses", {}, {}, "");
+  if (resp.statusCode != 200)
+  {
+    ignerr << "Failed to get license information from "
+      << _server.Url().Str() << "/" << _server.Version() << std::endl;
+  }
+  else if (!JSONParser::ParseLicenses(resp.data, this->dataPtr->licenses))
+  {
+    ignerr << "Failed to parse license information[" << resp.data << "]\n";
+  }
+}